+++ astropy/nddata/nduncertainty.py
@@ -262,6 +262,15 @@ class NDUncertainty(object):
     def parent_nddata(self, value):
         self._parent_nddata = value
 
+    def __repr__(self):
+        prefix = self.__class__.__name__ + '('
+        try:
+            body = np.array2string(self.array, separator=', ', prefix=prefix)
+        except AttributeError:
+            # In case it wasn't possible to use array2string
+            body = str(self.array)
+        return ''.join([prefix, body, ')'])
+
     def __getitem__(self, item):
         """
         Simple slicing is allowed and returns a reference *not* a copy. This
@@ -482,6 +491,32 @@ class StdDevUncertainty(NDUncertainty):
     Parameters
     ----------
     see `NDUncertainty`
+
+    Examples
+    --------
+    `StdDevUncertainty` should always be associated with an `NDData`-like
+    instance, either by creating it during initialization::
+
+        >>> from astropy.nddata import NDData, StdDevUncertainty
+        >>> ndd = NDData([1,2,3],
+        ...              uncertainty=StdDevUncertainty([0.1, 0.1, 0.1]))
+        >>> ndd.uncertainty
+        StdDevUncertainty([ 0.1,  0.1,  0.1])
+
+    or by setting it manually on the `NDData` instance::
+
+        >>> ndd.uncertainty = StdDevUncertainty([0.2], unit='m', copy=True)
+        >>> ndd.uncertainty
+        StdDevUncertainty([ 0.2])
+
+    the uncertainty ``array`` can also be set directly::
+
+        >>> ndd.uncertainty.array = 2
+        >>> ndd.uncertainty
+        StdDevUncertainty(2)
+
+    .. note::
+        The unit will not be displayed.
     """
 
     @property