+++ CHANGES.rst
@@ -196,6 +196,8 @@ API changes
   - ``NDUncertainty``: Public methods ``propagate_add``, etc. are replaced by
     a general ``propagate`` method. [#4272]
 
+  - ``NDUncertainty`` and subclasses: implement a representation (``__repr__``). [#4787]
+
   - ``StdDevUncertainty``: added an optional parameter ``copy`` which is False
     by default. [#4272]